requirements.txt
pandas==2.2.2
pyarrow==16.1.0
boto3==1.34.131
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
lightgbm==4.3.0
xgboost==2.0.3
scikit-learn==1.5.0
shap==0.45.1
great-expectations==0.18.19
mangum==0.17.0
